+++ src/server/physics/create-from-current-state.ts
@@ -35,6 +35,11 @@ export async function createPhysicsServerFromCurrentState() {
             ? {
                   connection: {
                       httpPort: Number(Opts.serverHttpPort),
+                      pingInterval: Opts.serverPingInterval,
+                      pingTimeout:
+                          Opts.serverPingTimeout == Opts.flatOpts.serverPingTimeout.max
+                              ? 10000e3
+                              : Opts.serverPingTimeout,
                       type: 'socket',
                   },
                   details: {